add first implementation of cross-compiling performance measurement
I'll add a README soon which explains how this works.

Right now, all you need to know is that you should
cd <repo>/apollo-harness
./scripts/run_test.sh

This may tell you to install some software (docker/podman, cross) and
then will run a test of the router bridge with a simple schema and
query.

The output is stored in results.