Add apollo-federation-error crate (#47)
For external tooling to continue to work with `federation-next` code, we
should generate error messages with the same code information.

This code adds the error codes, descriptions, and metadata in the
federation repo to the `apollo-federation-error` crate. Note that these
descriptions are not the error messages themselves, but rather the
description of the error code in auto-generated documentation. The
actual error message is determined by the call to `err()` (different but
related messages may be categorized under the same code). Wasn't exactly
sure how to embed error doc description and metadata into enum variants
via `thiserror` in a clean way, so there's a bit of indirection here.