SECOPS-2268: Add Gitleaks to CI (#49)
## Motivation / Implements

This PR adds the necessary config to configure
[Gitleaks](https://github.com/gitleaks/gitleaks) to run on PRs on this
repo. The Apollo Security team uses Gitleaks to test our repositories
for secrets.

Once this is accepted and merged, the Security team plans to make a
passing Gitleaks check a requirement for PRs to merge into this repo.
This will prevent secrets from being introduced to our repos.

In the event that a secret _is_ detected on a repo, the CI job will add
a comment to the PR associated with the detection to provide
instructions on how to properly resolve the detection. Additionally, if
a secret is detected, the Apollo Security team will be notified so that
we can be available to assist in resolving the detection.

If maintainers reviewing this PR have questions, please see this
[Apollo-internal Slack
link](https://apollograph.slack.com/archives/C03HCCJBAJC/p1696261536123059).

## Changed

- Updated `.circleci/config.yml` in this repo to contain appropriate
configuration to enable Gitleaks as a CI check.