Use single crate for integration tests (#96)
This PR changes our integration test layout to generate a single binary,
instead of having multiple (auto-discovered) binaries that are run
serially. This mimics the
[layout](https://github.com/apollographql/apollo-rs/blob/main/crates/apollo-compiler/Cargo.toml)
of `apollo-compiler`. This is also the recommendation of [the
docs](https://doc.rust-lang.org/cargo/reference/cargo-targets.html#integration-tests),
to increase test parallelism when there are many tests/decrease how long
it takes to run tests.

For reviewers, use "Hide Whitespace" when reviewing (this PR only
changes a few lines really).