feat: Inter-subgraph logic of federated query graph creation (#94)
This PR implements the inter-subgraph logic of federated query graph creation. That is, it implements the business logic for adding federated root nodes and cross-subgraph edges when creating a federated query graph. Note that this PR needs to use some functionality around operation processing eventually, but until that's ready, there are TODOs where the logic will be called. Tests will be added in follow-up PRs.