Be more diligent about `(Valid)FederationSchema`
For QP work, the API schema result needs to be a `FederationSchema`. We
also know that the API schema result is valid, so we make it a
`ValidFederationSchema`. API schema must also only be called on valid
schemas so the input is changed to a `ValidFederationSchema`.

Supergraphs are also federation schemas so this changes the constructors
to build a `ValidFederationSchema` early.

`ValidFederationSchema` contains an `Arc` while `FederationSchema`
doesn't, and that is relied upon in a few places. I add some clones of
the inner schema to convert from an immutable `ValidFederationSchema` to
a mutable `FederationSchema`, it maybe could be done more efficiently,
but it's OK.