fix: more compatible definitions added by `include_{defer,stream}` (#221)
There was a rather silly mistake in the `@defer` definition: it applies
to fragment spreads, but I copy-pasted wrong and made it apply to fields
instead. That's fixed now.

I also removed the doc comments from the definitions. The JS code did
not add them and having them messes up the comparison between JS and
Rust results in the router. If it's any use to have it, we can re-add it
later when the JS implementation is gone.