feat: support fed v2 `@core` links
This is for backward compatibility with existing supergraphs, bolting
`@core` support onto the `@link` reading. It does not do anything to
support fed v1 supergraphs.

While `@core` is not used in new graphs today, it is supported by the JS
code, and we have a lot of tests throughout our systems that rely on
using `@core` with fed v2.