Fix nontrivial conditions selection set comparison
I think this must only compare the fields that are selected, not the
parent type and the schema it's a part of. That's because the selections
are parsed in the context of a subgraph schema, so writing the same
`@key()` in subgraph A and B results in selection sets that don't
compare equal, but their actual selections do. And that is what key
resolution cares about.