feat: support legacy `@core` link syntax (#223)
This is for backward compatibility with existing supergraphs, bolting
`@core` support onto the `@link` reading. It does not do anything to
support fed v1 supergraphs.

While `@core` is not used in new graphs today, it is supported by the JS
code, and we have a lot of tests throughout our systems that rely on
using `@core` with fed v2.

Additionally, because the implementation of API schema generation is the
same between fed v1 and fed v2, supporting `@core` will let us land it
in the router *without* a fallback to JS for fed v1 supergraphs.

All new code that supports `@core` mentions `@core` literally in a
comment so we can rip it out easily when we migrate over customers and
our tests in the router.