Use boxes inside query plan nodes
I don't think these need to be Arc. The parallel structures aren't Arc in the router today, and we don't need them to be internally. A few boxes to keep the size of the structure finite (and even small) should be enough.

But the main point of this change is to make TopLevelPlanNode and PlanNode
have the same shape for shared branches, so we can simply copy the node
values around to convert a PlanNode to a TopLevelPlanNode.