Add skeleton CLI tool (#246)
Example usage: `cargo cli api
../router/examples/graphql/supergraph.graphql`

This executable is intended to be the successor of
`composer-tool`. Initially it has a
single `api` subcommand that converts a supergraph schema to its
corresponding API schema. (Error handling can be improved.)

The CLI is a separate crate so that the library crate does not need to
depend on `clap`. See https://blog.axo.dev/2024/03/its-a-lib-and-a-bin