Add some Debug impls
The derived Debug impl for `GraphPath` included the whole query graph,
which made it untenable to read. This patch removes that field from the
Debug representation.

OpGraphPaths have a terse string representation. In other structures
that contain OpGraphPaths, I think the terse representation is
preferable. This adds Debug impls to use the terse representation in
those cases.