Ported anImplementationIsEntityWithFieldShareable from JS (#260)
This PR addresses FED-33 and is a port of
[`anImplementationIsEntityWithFieldShareable`](https://github.com/apollographql/federation/blob/main/query-graphs-js/src/graphPath.ts#L2279).

There are a number of differences between how the JS code and Rust code
represent query graphs, graph paths, nodes, and edges. The original
source code is linked above. I tried to keep everything one-to-one, but
there is one thing that seems to not be needed: the `baseType` function.
This is called a few times in the original JS code, but I was unable to
find a translation for it.

---------

Co-authored-by: Renée <renee.kooi@apollographql.com>