Make operation modules more-internal
The `operation::normalized_*` set of modules exist for internal privacy
reasons. This detail does not need to be exposed from the `operation`
module. This change makes the internal modules fully private and exports
the types the modules provide from `operation` directly. (An exception
is `normalized_selection_map`, which is public so its `Entry` type can
be accessed, similarly to how std has `collections::HashMap` and
`collections::hash_map::Entry`).